OLDFIles/influxDb/requirements.txt
fastapi
uvicorn
influxdb-client
python-dotenv
pyserial
